use cabal sandbox for travis builds, HEAD_DEPS
This avoids a number of problems with dependency resolution.

We want to install dependencies in a different travis phase than
building the package under test, so that travis can distinguish upstream
failures from local failures.  The dependencies include both HEAD_DEPS
and packages which are not dependencies of any HEAD_DEPS.

We used to invoke cabal once for the HEAD_DEPS, and once for the current
package with `--only-dependencies`.  This leads to diamond dependency
problems, which are often masked by cabal building the current package
with diagrams-* versions from Hackage, instead of the HEAD_DEPS.  For
example, cabal tries to rebuild `diagrams-core` with a different version
of `profunctors`, but doesn't know to pick diagrams-core HEAD.

More recently, we call `cabal install --only-dependencies` with both the
HEAD_DEPS and the current package.  This does seem to help with some
diamond dependency issues, but cabal gives up when one HEAD_DEP depends
on another.  For example `cabal install --only-dependencies
monoid-extras/ diagrams-core/ .` asks cabal to install `monoid-extras`
and not to install `monoid-extras`.

In the context of a cabal sandbox, cabal's solver behaves somewhat
differently.  It will only ever install one version of one package, and
we can register HEAD_DEPS without installing anything.  For example,
`cabal add-source diagrams-core` implies that cabal should not install
any version of `diagrams-core` except HEAD.  This should prevent both
problems outlined above.  We have only one command to install the
HEAD_DEPS and all dependencies, so cabal resolves all version
constraints together.